ffi: remove eval.rs
