+++ pyo3-ffi/src/ceval.rs
@@ -3,6 +3,27 @@ use crate::pystate::PyThreadState;
 use std::os::raw::{c_char, c_int, c_void};
 
 extern "C" {
+    #[cfg_attr(PyPy, link_name = "PyPyEval_EvalCode")]
+    pub fn PyEval_EvalCode(
+        arg1: *mut PyObject,
+        arg2: *mut PyObject,
+        arg3: *mut PyObject,
+    ) -> *mut PyObject;
+
+    pub fn PyEval_EvalCodeEx(
+        co: *mut PyObject,
+        globals: *mut PyObject,
+        locals: *mut PyObject,
+        args: *const *mut PyObject,
+        argc: c_int,
+        kwds: *const *mut PyObject,
+        kwdc: c_int,
+        defs: *const *mut PyObject,
+        defc: c_int,
+        kwdefs: *mut PyObject,
+        closure: *mut PyObject,
+    ) -> *mut PyObject;
+
     #[cfg_attr(Py_3_9, deprecated(note = "Python 3.9"))]
     #[cfg_attr(PyPy, link_name = "PyPyEval_CallObjectWithKeywords")]
     pub fn PyEval_CallObjectWithKeywords(
